requirements.txt
streamlit
langchain
langchain-groq
pyttsx3
python-docx
PyPDF2
openpyxl
python-dotenv